move algoliasearch code to right folder and file
